% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmfr.R
\name{a_gmfr}
\alias{a_gmfr}
\title{a_gmfr}
\usage{
a_gmfr(x, y, na.rm = F)
}
\arguments{
\item{x}{a vector of numeric values}

\item{y}{a vector of numeric values}

\item{na.rm}{boolean flag indicating whether or not to remove NA values from
computation}
}
\value{
a numeric value
}
\description{
Get the intercept for the geometric mean functional relationship model.
}
