% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac.R
\name{ac}
\alias{ac}
\title{ac}
\usage{
ac(x, y, na.rm = F)
}
\arguments{
\item{x}{a vector of numeric values}

\item{y}{a vector of numeric values}

\item{na.rm}{boolean flag indicating whether or not to remove NA values from
computation}
}
\value{
a numeric value bounded \link{0, 1}
}
\description{
Compute agreement coefficient (AC) following
\href{https://www.ingentaconnect.com/content/asprs/pers/2006/00000072/00000007/art00006}{Ji and Gallo 2006}
}
\details{
AC is bounded between 0 and 1 where 1 represents perfect agreement and 0
represents no agreement.
}
