% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{agg_extraction}
\alias{agg_extraction}
\title{Aggregate Unit Extraction}
\usage{
agg_extraction(
  surface,
  plots,
  agg_columns,
  agg_units,
  id = "PLOT",
  sampling_correction = TRUE,
  na.rm = FALSE
)
}
\arguments{
\item{surface}{Raster::RasterLayer or terra::SpatRaster}

\item{plots}{sf::sf or terra::SpatVector}

\item{agg_units}{sf::sf polygons demarking units over which comparisons
will be made}

\item{id}{string indicating of column that identifies individual plots}

\item{sampling_correction}{boolean. If TRUE only pixels coincident with
plot polygons will be extracted. If FALSE all pixels with an agg_unit
will be extracted for copmarison.}

\item{na.rm}{boolean indicating wether or not NA pixels should be ignored
in the extraction}
}
\value{
sf::sf object containing aggregation unit polygons and summarized
values
}
\description{
Extract map values within aggregation units for comparison
}
