% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{agg_summary}
\alias{agg_summary}
\title{Aggregate summary}
\usage{
agg_summary(
  surface,
  plots,
  y,
  agg_units,
  id = "PLOT",
  sampling_correction = TRUE,
  na.rm = FALSE
)
}
\arguments{
\item{surface}{RasterLayer or SpatRaster containing predictions}

\item{plots}{sf::sf object containing plot polygons of reference data}

\item{y}{name of response contained in plots}

\item{agg_units}{sf::sf polygons demarking units over which comparisons
will be made}

\item{id}{name of field uniquely identifying plots}

\item{sampling_correction}{boolean. If TRUE only pixels coincident with
plot polygons will be extracted. If FALSE all pixels with an agg_unit
will be extracted for copmarison.}

\item{na.rm}{boolean indicating whether or not to ignore NA values}
}
\value{
a data.frame
}
\description{
Compute agreement summary over aggregation units
}
