% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{aggregate_plot_pixel}
\alias{aggregate_plot_pixel}
\title{Aggregate plot pixel}
\usage{
aggregate_plot_pixel(plots, agg_units, x, y, na.rm = FALSE)
}
\arguments{
\item{plots}{results from plot_pixel_extraction}

\item{agg_units}{sf::sf polygons demarking units over which comparisons
will be made}

\item{x}{modeled value}

\item{y}{reference value}

\item{na.rm}{boolean indicating wether or not NA pixels should be ignored
in the extraction}
}
\value{
sf::sf object containing aggregation unit polygons and summarized
values
}
\description{
Aggregate tabular results from plot_pixel_extraction
}
