% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{compute_plot_centroids}
\alias{compute_plot_centroids}
\title{Compute plot centroids}
\usage{
compute_plot_centroids(plots, id = "PLOT", agg_columns = NULL, na.rm = FALSE)
}
\arguments{
\item{plots}{sf::sf or terra::SpatVector}

\item{id}{string indicating of column that identifies individual plots}
}
\value{
sf::sf object containing point geometries
}
\description{
Compute plot centroids
}
