% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data}
\format{
A data.table with 344 rows and 2 variables:
\describe{
\item{observed}{observed value}
\item{predicted}{model predicted value}
...
}
}
\usage{
example_data
}
\description{
A dataset containing the observed and predicted values for 344 observations
used to develop an anonymous model
}
\keyword{datasets}
