% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmfr.R
\name{gmfr}
\alias{gmfr}
\title{gmfr}
\usage{
gmfr(x, y, na.rm = F)
}
\arguments{
\item{x}{a vector of numeric values}

\item{y}{a vector of numeric values}

\item{na.rm}{boolean flag indicating whether or not to remove NA values from
computation}
}
\value{
a list of vectors (x, and y)
}
\description{
Computes x and y coordinates along the GMFR line for input vectors x and y
}
