% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tesselationR.R
\name{hex_tesselation}
\alias{hex_tesselation}
\title{Hexagonal Tesselation}
\usage{
hex_tesselation(separation, bounds)
}
\arguments{
\item{separation}{distance between hexagonal centroids in km}

\item{bounds}{a bounding shape (sf::sf object) in a projected coordinate
system with units in meters}
}
\value{
an sf::sf object
}
\description{
Produce a hexaganal tesselation across a shape
}
