% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac.R
\name{mpds}
\alias{mpds}
\title{mpds}
\usage{
mpds(data, x, y, na.rm = F)
}
\arguments{
\item{data}{data.frame}

\item{x}{column name (numeric)}

\item{y}{column name (numeric)}

\item{na.rm}{boolean flag indicating whether or not to remove NA values from
computation}
}
\value{
numeric value
}
\description{
Systematic mean product-difference
}
