% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{p_mae}
\alias{p_mae}
\title{Percent mae}
\usage{
p_mae(data, x, y, na.rm = FALSE)
}
\arguments{
\item{data}{data.frame}

\item{x}{column name for predicted data}

\item{y}{column name for response data}

\item{na.rm}{boolean flag indicating whether or not to remove NA values from
computation}
}
\value{

}
\description{
mae divided by mean response value
}
