% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{plot_pixel_extraction}
\alias{plot_pixel_extraction}
\title{Plot Pixel Extraction}
\usage{
plot_pixel_extraction(
  surface,
  plots,
  agg_columns = NULL,
  id = "PLOT",
  na.rm = FALSE
)
}
\arguments{
\item{surface}{Raster::RasterLayer or terra::SpatRaster}

\item{plots}{sf::sf or terra::SpatVector}

\item{id}{string indicating of column that identifies individual plots}

\item{na.rm}{boolean indicating wether or not NA pixels should be ignored
in the extraction}
}
\value{
an sf object with val column added
}
\description{
Extract map values at plot polygons for comparison
}
