% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{plot_pixel_summary}
\alias{plot_pixel_summary}
\title{Plot pixel summary}
\usage{
plot_pixel_summary(
  surface,
  plots,
  y,
  agg_columns = NULL,
  id = "PLOT",
  na.rm = FALSE
)
}
\arguments{
\item{surface}{RasterLayer or SpatRaster containing predictions}

\item{plots}{sf::sf object containing plot polygons of reference data}

\item{y}{name of response contained in plots}

\item{agg_columns}{vector of column names that we want to aggregate over}

\item{id}{name of field uniquely identifying plots}

\item{na.rm}{boolean indicating whether or not to ignore NA values}
}
\value{
a data.frame
}
\description{
Compute agreement summary over plots
}
