% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rmse}
\alias{rmse}
\title{rmse}
\usage{
rmse(data, x, y, na.rm = FALSE)
}
\arguments{
\item{data}{data.frame}

\item{x}{column name for predicted data}

\item{y}{column name for response data}

\item{na.rm}{boolean flag indicating whether or not to remove NA values from
computation}
}
\value{
a numeric value
}
\description{
compute the root mean squared error
}
