% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spod.R
\name{spod}
\alias{spod}
\title{spod}
\usage{
spod(x, y, na.rm = F)
}
\arguments{
\item{x}{a vector of numeric values}

\item{y}{a vector of numeric values}

\item{na.rm}{boolean flag indicating whether or not to remove NA values from
computation}
}
\value{
numeric value
}
\description{
Sum of potential difference
}
