% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_table}
\alias{summary_table}
\title{Summary agreement metrics}
\usage{
summary_table(data, x, y, na.rm = FALSE)
}
\arguments{
\item{data}{a data.frame containing two datasets to compare (x and y)}

\item{x}{name of column in data (predicted)}

\item{y}{name of column in data (response)}

\item{na.rm}{boolean indicating whether or not to ignore NA values}
}
\value{
a data.frame
}
\description{
Summary agreement metrics
}
