% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vec_narm}
\alias{vec_narm}
\title{vec_narm}
\usage{
vec_narm(x, y)
}
\arguments{
\item{x}{a vector of numeric values}

\item{y}{a vector of numeric values}
}
\value{
a vector containing two vectors: c(x, y)
}
\description{
Helper to remove na vales from vectors of matching length
}
