% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmfr.R
\name{gmfr}
\alias{gmfr}
\title{gmfr}
\usage{
gmfr(data, x, y, na.rm = F)
}
\arguments{
\item{data}{data.frame}

\item{x}{column name for independent data in model}

\item{y}{column name for response data in model}

\item{na.rm}{boolean flag indicating whether or not to remove NA values from
computation}
}
\value{
a list of vectors (x, and y)
}
\description{
Computes x and y coordinates along the GMFR line for input vectors x and y
}
