% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac.R
\name{msd}
\alias{msd}
\title{msd}
\usage{
msd(data, x, y, na.rm = F)
}
\arguments{
\item{data}{data.frame}

\item{x}{column name for independent data in model}

\item{y}{column name for response data in model}

\item{na.rm}{boolean flag indicating whether or not to remove NA values from
computation}
}
\value{
numeric value
}
\description{
Mean squared difference
}
