% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac.R
\name{spds}
\alias{spds}
\title{spds}
\usage{
spds(x, y, na.rm = F)
}
\arguments{
\item{x}{a vector of numeric values}

\item{y}{a vector of numeric values}

\item{na.rm}{boolean flag indicating whether or not to remove NA values from
computation}
}
\value{
numeric value
}
\description{
Systematic sum of product-difference
}
